+++ nle_code_wrapper/bot/strategies/__init__.py
@@ -1,13 +1,13 @@
-from nle_code_wrapper.bot.strategies.exploration_strategies import (
+from nle_code_wrapper.bot.strategies.explore import (
     explore_corridor,
     explore_room,
     explore_room_systematically,
+    general_explore,
     goto_next_unexplored_room,
     search_corridor_for_hidden_doors,
     search_for_traps,
     search_room_for_hidden_doors,
 )
-from nle_code_wrapper.bot.strategies.explore import explore
 from nle_code_wrapper.bot.strategies.explore_items import explore_items
 from nle_code_wrapper.bot.strategies.fight_monster import fight_all_monsters, fight_closest_monster
 from nle_code_wrapper.bot.strategies.goto import goto
@@ -21,7 +21,7 @@ from nle_code_wrapper.bot.strategies.run_away import run_away
 from nle_code_wrapper.bot.strategies.search import search
 
 __all__ = [
-    explore,
+    general_explore,
     search,
     explore_items,
     goto_stairs,
@@ -36,11 +36,4 @@ __all__ = [
     goto_items,
     quaff_potion_from_inv,
     goto,
-    search_for_traps,
-    explore_room_systematically,
-    explore_room,
-    goto_next_unexplored_room,
-    explore_corridor,
-    search_corridor_for_hidden_doors,
-    search_room_for_hidden_doors,
 ]
