+++ tests/minihack/test_fight_corridor.py
@@ -3,7 +3,7 @@ from nle_utils.play import play
 
 from nle_code_wrapper.bot.bot import Bot
 from nle_code_wrapper.bot.exceptions import BotPanic
-from nle_code_wrapper.bot.strategies import explore, fight_closest_monster, goto_stairs, run_away
+from nle_code_wrapper.bot.strategies import fight_closest_monster, general_explore, goto_stairs, run_away
 from nle_code_wrapper.envs.minihack.play_minihack import parse_minihack_args
 
 
@@ -30,7 +30,7 @@ class TestMazewalkMapped(object):
                     elif goto_stairs(bot):
                         pass
                     else:
-                        explore(bot)
+                        general_explore(bot)
                 except BotPanic:
                     pass
