+++ tests/minihack/test_lava_crossing.py
@@ -3,7 +3,7 @@ from nle_utils.play import play
 
 from nle_code_wrapper.bot.bot import Bot
 from nle_code_wrapper.bot.exceptions import BotPanic
-from nle_code_wrapper.bot.strategies import explore, goto_stairs
+from nle_code_wrapper.bot.strategies import general_explore, goto_stairs
 from nle_code_wrapper.envs.minihack.play_minihack import parse_minihack_args
 
 
@@ -18,16 +18,16 @@ class TestMazewalkMapped(object):
     def test_solve_lava_crossing(self, env):
         cfg = parse_minihack_args(argv=[f"--env={env}", "--no-render"])
 
-        def general_explore(bot: "Bot"):
+        def solve(bot: "Bot"):
             while True:
                 try:
                     if goto_stairs(bot):
                         pass
                     else:
-                        explore(bot)
+                        general_explore(bot)
                 except BotPanic:
                     pass
 
-        cfg.strategies = [general_explore]
+        cfg.strategies = [solve]
         status = play(cfg)
         assert status["end_status"].name == "TASK_SUCCESSFUL"
