+++ tests/minihack/test_multiroom.py
@@ -3,7 +3,7 @@ from nle_utils.play import play
 
 from nle_code_wrapper.bot.bot import Bot
 from nle_code_wrapper.bot.exceptions import BotPanic
-from nle_code_wrapper.bot.strategies import explore, fight_closest_monster, goto_stairs, open_doors, run_away
+from nle_code_wrapper.bot.strategies import fight_closest_monster, general_explore, goto_stairs, open_doors, run_away
 from nle_code_wrapper.envs.minihack.play_minihack import parse_minihack_args
 
 
@@ -37,7 +37,7 @@ class TestMazewalkMapped(object):
                     elif open_doors(bot):
                         pass
                     else:
-                        explore(bot)
+                        general_explore(bot)
                 except BotPanic:
                     pass
