+++ tests/minihack/test_room.py
@@ -3,7 +3,7 @@ from nle_utils.play import play
 
 from nle_code_wrapper.bot.bot import Bot
 from nle_code_wrapper.bot.exceptions import BotPanic
-from nle_code_wrapper.bot.strategies import explore, fight_closest_monster, goto_stairs, run_away
+from nle_code_wrapper.bot.strategies import fight_closest_monster, general_explore, goto_stairs, run_away
 from nle_code_wrapper.envs.minihack.play_minihack import parse_minihack_args
 
 
@@ -23,17 +23,17 @@ class TestMazewalkMapped(object):
     def test_solve_room_explore(self, env):
         cfg = parse_minihack_args(argv=[f"--env={env}", "--no-render"])
 
-        def general_explore(bot: "Bot"):
+        def solve(bot: "Bot"):
             while True:
                 try:
                     if goto_stairs(bot):
                         pass
                     else:
-                        explore(bot)
+                        general_explore(bot)
                 except BotPanic:
                     pass
 
-        cfg.strategies = [general_explore]
+        cfg.strategies = [solve]
         status = play(cfg)
         assert status["end_status"].name == "TASK_SUCCESSFUL"
 
@@ -56,7 +56,7 @@ class TestMazewalkMapped(object):
                     elif goto_stairs(bot):
                         pass
                     else:
-                        explore(bot)
+                        general_explore(bot)
                 except BotPanic:
                     pass
 
@@ -80,7 +80,7 @@ class TestMazewalkMapped(object):
                     if goto_stairs(bot):
                         pass
                     else:
-                        explore(bot)
+                        general_explore(bot)
                 except BotPanic:
                     pass
 
@@ -108,7 +108,7 @@ class TestMazewalkMapped(object):
                     elif goto_stairs(bot):
                         pass
                     else:
-                        explore(bot)
+                        general_explore(bot)
                 except BotPanic:
                     pass
